+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ## Unreleased
 - Fix issue with colormap clamping past max value
 - Improve performance for large datasets by batching into fewer threejs objects
-- Add jump flood outline rendering for selected object
+- Add jump flood outline rendering for selected object/face
+- Allow face picking outside of semantics mode
 
 ## 0.6.0 - 2026-05-08